refactor: 자동 로그인 로직 구조 변경
